refactor: 가독성을 높이기 위해 전체 디렉토리 구조 개편
