+++ src/event/initEvent.js
@@ -1,5 +1,5 @@
-import selectMenuButtonEvent from './selectMenuButtonEvent.js';
-import addStationEvent from './addStationEvent.js';
+import selectMenuButtonEvent from './common/selectMenuButtonEvent.js';
+import addStationEvent from './station/addStationEvent.js';
 
 export default function initEvent() {
   selectMenuButtonEvent();
