+++ src/render/render.js
@@ -1,5 +1,5 @@
 import renderSelector from './line/renderSelector.js';
-import renderStationList from './renderStationList.js';
+import renderStationList from './station/renderStationList.js';
 
 function clearInputValue() {
   const $input = document.querySelectorAll('input');
